requirements.txt
langchain==1.0.7
langchain_classic==1.0.0
langchain_community==0.4.1
langchain_core==1.0.5
langchain_google_community==3.0.1
langchain_huggingface==1.0.1
langchain_openai==1.0.3
langchain_qdrant==1.1.0
langchain_text_splitters==1.0.0
langgraph==1.0.3
langsmith==0.4.43
pydantic==2.12.4
python-dotenv==1.2.1
qdrant_client==1.16.0
streamlit==1.51.0
typing_extensions==4.15.0
